Move setMini to pokemoninfo
