+++ src/ui/battle-info/player-battle-info.ts
@@ -1,6 +1,7 @@
-import { getLevelTotalExp } from "#app/data/exp";
+import { getLevelRelExp, getLevelTotalExp } from "#app/data/exp";
 import type { PlayerPokemon } from "#app/field/pokemon";
 import { globalScene } from "#app/global-scene";
+import { ExpGainsSpeed } from "#enums/exp-gains-speed";
 import { Stat } from "#enums/stat";
 import BattleInfo from "./battle-info";
 
@@ -42,5 +43,112 @@ export class PlayerBattleInfo extends BattleInfo {
     this.lastLevelExp = pokemon.levelExp;
 
     this.statValuesContainer.setPosition(8, 7);
+  }
+
+  override async updatePokemonExp(
+    pokemon: PlayerPokemon,
+    instant?: boolean,
+    levelDurationMultiplier = 1,
+  ): Promise<void> {
+    const levelUp = this.lastLevel < pokemon.level;
+    const relLevelExp = getLevelRelExp(this.lastLevel + 1, pokemon.species.growthRate);
+    const levelExp = levelUp ? relLevelExp : pokemon.levelExp;
+    let ratio = relLevelExp ? levelExp / relLevelExp : 0;
+    if (this.lastLevel >= globalScene.getMaxExpLevel(true)) {
+      if (levelUp) {
+        ratio = 1;
+      } else {
+        ratio = 0;
+      }
+      instant = true;
+    }
+    const durationMultiplier = Phaser.Tweens.Builders.GetEaseFunction("Sine.easeIn")(
+      1 - Math.max(this.lastLevel - 100, 0) / 150,
+    );
+    let duration =
+      this.visible && !instant
+        ? ((levelExp - this.lastLevelExp) / relLevelExp) *
+          BattleInfo.EXP_GAINS_DURATION_BASE *
+          durationMultiplier *
+          levelDurationMultiplier
+        : 0;
+    const speed = globalScene.expGainsSpeed;
+    if (speed && speed >= ExpGainsSpeed.DEFAULT) {
+      duration = speed >= ExpGainsSpeed.SKIP ? ExpGainsSpeed.DEFAULT : duration / Math.pow(2, speed);
+    }
+    if (ratio === 1) {
+      this.lastLevelExp = 0;
+      this.lastLevel++;
+    } else {
+      this.lastExp = pokemon.exp;
+      this.lastLevelExp = pokemon.levelExp;
+    }
+    if (duration) {
+      globalScene.playSound("se/exp");
+    }
+    return new Promise(resolve => {
+      globalScene.tweens.add({
+        targets: this.expMaskRect,
+        ease: "Sine.easeIn",
+        x: ratio * 510,
+        duration: duration,
+        onComplete: () => {
+          if (!globalScene) {
+            return resolve();
+          }
+          if (duration) {
+            globalScene.sound.stopByKey("se/exp");
+          }
+          if (ratio === 1) {
+            globalScene.playSound("se/level_up");
+            this.setLevel(this.lastLevel);
+            globalScene.time.delayedCall(500 * levelDurationMultiplier, () => {
+              this.expMaskRect.x = 0;
+              this.updateInfo(pokemon, instant).then(() => resolve());
+            });
+            return;
+          }
+          resolve();
+        },
+      });
+    });
+  }
+
+  setMini(mini: boolean): void {
+    if (this.mini === mini) {
+      return;
+    }
+
+    this.mini = mini;
+
+    this.box.setTexture(this.getTextureName());
+    this.statsBox.setTexture(`${this.getTextureName()}_stats`);
+
+    if (this.player) {
+      this.y -= 12 * (mini ? 1 : -1);
+      this.baseY = this.y;
+    }
+
+    const offsetElements = [
+      this.nameText,
+      this.genderText,
+      this.teraIcon,
+      this.splicedIcon,
+      this.shinyIcon,
+      this.statusIndicator,
+      this.levelContainer,
+    ];
+    offsetElements.forEach(el => (el.y += 1.5 * (mini ? -1 : 1)));
+
+    [this.type1Icon, this.type2Icon, this.type3Icon].forEach(el => {
+      el.x += 4 * (mini ? 1 : -1);
+      el.y += -8 * (mini ? 1 : -1);
+    });
+
+    this.statValuesContainer.x += 2 * (mini ? 1 : -1);
+    this.statValuesContainer.y += -7 * (mini ? 1 : -1);
+
+    const toggledElements = [this.hpNumbersContainer, this.expBar];
+    toggledElements.forEach(el => el.setVisible(!mini));
   }
 }
